+++ resources/views/pages/partials/blocks/container.blade.php
@@ -1,5 +1,5 @@
 @php($class = collect(['wb-container', $block->containerWidthClass(), 'wb-stack'])->filter()->implode(' '))
-<div class="{{ $class }}">
+<div class="{{ $class }}" data-wb-public-block-type="{{ $block->publicBlockTypeAttribute() }}">
     @foreach ($block->children as $child)
         @include('pages.partials.block', ['block' => $child])
     @endforeach